Recover the gameplay layer, and split CLAUDE.md into per-subsystem records
Six notes files recovered from the binary against the shipped manual, which
documents several features the retail build cannot execute:

- `ai_behaviour_notes.md` - the vision cone is a **swept ray**, detecting only
  on the tick it crosses a target's bearing; and the orange "reacquiring" alert
  state is unreachable by any path in the binary.
- `combat_notes.md` - there is **no hitscan**; every shot builds a
  `ProjectileActor`. Armour is a flat absorb threshold, splash has no falloff.
- `orders_notes.md` - the `PendingOrder` FIFO and its 11 kinds, of which
  nothing on the client ever pushes one. **Formations have no geometry.**
- `navigation_notes.md` - flat A* with *squared* distance as both cost and
  heuristic, so paths are non-optimal, and a node budget that emits a partial
  path.
- `inventory_notes.md`, `stealth_and_fog_notes.md`, `gadgets_notes.md`.

CLAUDE.md was 194 KB of context before anyone had asked a question. It is now
an index: the address map and each subsystem's design record live in their own
file, and only the rules that govern *writing* the code stay behind.

Ghidra work moves into a `ghidra-analyst` subagent, so decompiler output stays
out of the main context, with a hook holding the line.

+++ src/MakeRole.cpp
@@ -68,9 +68,8 @@ float RevolutionsToBam(double revolutions) {
   return static_cast<float>(revolutions * BamPerTurn);
 }
 
-float ToFixed16(double cycles_per_second) {
-  return static_cast<float>(
-      RoundToNearest(static_cast<float>(cycles_per_second) * FixedScale));
+int ToFixed16(double cycles_per_second) {
+  return RoundToNearest(static_cast<float>(cycles_per_second) * FixedScale);
 }
 
 Character *MakeCharacter(const CharacterDesc &d) {
@@ -138,10 +137,8 @@ Character *MakeCharacter(const CharacterDesc &d) {
   // - reproducing that is the difference between matching the game and being
   // half an ulp out on every turning character.
   if (c->turning_speed > 0.0f && c->size > 0.0f) {
-    float cycles = static_cast<float>(static_cast<int>(c->walking_speed)) /
-                   FixedScale;
-    c->walking_speed =
-        static_cast<float>(RoundToNearest((cycles / c->size) * FixedScale));
+    float cycles = static_cast<float>(c->walking_speed) / FixedScale;
+    c->walking_speed = RoundToNearest((cycles / c->size) * FixedScale);
   }
   // A max below the initial is clamped up, not down.
   if (c->maximum_first_person_range <= c->initial_first_person_range &&
